# bffs/reisefrihet-fastify/contract/mobile.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wayfare reisefrihet mobile API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/travel-options": {
      "get": {
        "tags": ["travel"],
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "fareFrameId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["passengerProfiles", "luggage"],
                  "properties": {
                    "passengerProfiles": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["id", "label", "category", "required"],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "category": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": ["passenger"]
                              },
                              {
                                "type": "string",
                                "enum": ["luggage"]
                              }
                            ]
                          },
                          "required": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "luggage": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["id", "label", "category", "required"],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "category": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": ["passenger"]
                              },
                              {
                                "type": "string",
                                "enum": ["luggage"]
                              }
                            ]
                          },
                          "required": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/journeys/check-in": {
      "post": {
        "tags": ["journeys"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customerNumber", "userProfileIds", "luggageIds"],
                "properties": {
                  "customerNumber": {
                    "pattern": "^[0-9]+$",
                    "minLength": 1,
                    "maxLength": 15,
                    "description": "Customer number represented as digits to avoid JSON precision loss",
                    "type": "string"
                  },
                  "userProfileIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "luggageIds": {
                    "default": [],
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["journeyId", "startTime"],
                  "properties": {
                    "journeyId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "startTime": {
                      "format": "date-time",
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/journeys/{journeyId}/check-out": {
      "post": {
        "tags": ["journeys"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customerNumber", "startTime"],
                "properties": {
                  "customerNumber": {
                    "pattern": "^[0-9]+$",
                    "minLength": 1,
                    "maxLength": 15,
                    "description": "Customer number represented as digits to avoid JSON precision loss",
                    "type": "string"
                  },
                  "startTime": {
                    "format": "date-time",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "journeyId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "startTime",
                    "endTime",
                    "totalCost",
                    "currency",
                    "modes",
                    "passengers",
                    "legs",
                    "zones",
                    "simulated"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "startTime": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "endTime": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "totalCost": {
                      "type": "number"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "modes": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "passengers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": ["profile", "label"],
                        "properties": {
                          "profile": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "legs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "mode",
                          "from",
                          "to",
                          "startTime",
                          "endTime",
                          "zones"
                        ],
                        "properties": {
                          "mode": {
                            "type": "string"
                          },
                          "from": {
                            "type": "string"
                          },
                          "to": {
                            "type": "string"
                          },
                          "startTime": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "endTime": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "distanceMeters": {
                            "minimum": 0,
                            "type": "integer"
                          },
                          "zones": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "zones": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "confidence": {
                      "type": "number"
                    },
                    "simulated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/customers/{customerNumber}/payment-status": {
      "get": {
        "tags": ["payments"],
        "parameters": [
          {
            "schema": {
              "pattern": "^[0-9]+$",
              "minLength": 1,
              "maxLength": 15,
              "type": "string"
            },
            "in": "path",
            "name": "customerNumber",
            "required": true,
            "description": "Customer number represented as digits to avoid JSON precision loss"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["hasFailedTransactions"],
                  "properties": {
                    "hasFailedTransactions": {
                      "type": "boolean"
                    },
                    "overdueAmount": {
                      "type": "number"
                    },
                    "currency": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "504": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["code", "message", "retryable"],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "retryable": {
                      "type": "boolean"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}